move: ShopIngredient to shopping_list
