+++ app/src/main/java/com/androidimpact/app/fragments/ShoppingListFragment.java
@@ -18,22 +18,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.CompoundButton;
 import android.widget.Spinner;
 import android.widget.Switch;
 import android.widget.TextView;
 
-import com.androidimpact.app.activities.AddEditStoreIngredientActivity;
-import com.androidimpact.app.activities.MainActivity;
-import com.androidimpact.app.activities.RecipeAddViewEditActivity;
-import com.androidimpact.app.ingredients.Ingredient;
 import com.androidimpact.app.R;
-import com.androidimpact.app.ingredients.ShopIngredient;
-import com.androidimpact.app.ingredients.StoreIngredient;
+import com.androidimpact.app.shopping_list.ShopIngredient;
 import com.androidimpact.app.shopping_list.AddShoppingListItemActivity;
 import com.androidimpact.app.shopping_list.ShopIngredientAdapter;
 import com.androidimpact.app.shopping_list.ShoppingList;
-import com.androidimpact.app.ingredients.StoreIngredientViewAdapter;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.firestore.CollectionReference;
